refactor(reader): lift Volume/ReaderState collaborators to commonMain (#953)
Move the shared reader infra the CBZ ViewModel needs — VolumeNavEvent,
VolumeNavigationController, VolumeKeyDispatcher, ReaderStateHolder — from :app to
:feature:reader/commonMain (all pure Kotlin, no Android deps). :app keeps typealias
shims (the #946 idiom) so epub/pdf reader consumers are unchanged, except VolumeNavEvent
consumers import the new package directly (a typealias can't expose a sealed class's
nested objects like VolumeNavEvent.Forward).

Enables the CBZ ViewModel to reference these from commonMain in the upcoming move.

Progresses #953.
