+++ app/src/main/kotlin/com/riffle/app/MainActivity.kt
@@ -30,7 +30,7 @@ import com.riffle.app.feature.reader.ReaderStateHolder
 import com.riffle.app.i18n.AppLocaleController
 import com.riffle.app.feature.reader.VolumeKeyAction
 import com.riffle.app.feature.reader.VolumeKeyEventHandler
-import com.riffle.app.feature.reader.VolumeNavEvent
+import com.riffle.feature.reader.VolumeNavEvent
 import com.riffle.app.feature.reader.VolumeNavigationController
 import com.riffle.app.feature.reader.autoscroll.AutoScrollController
 import com.riffle.core.domain.autoscroll.AutoScrollEvent
