+++ app/src/main/kotlin/com/riffle/app/feature/reader/ReaderStateHolder.kt
@@ -1,12 +1,4 @@
 package com.riffle.app.feature.reader
 
-
-
-class ReaderStateHolder constructor() {
-    @Volatile var isReaderActive: Boolean = false
-    @Volatile var isPanelOpen: Boolean = false
-
-    // True while in-app audio (Readaloud today; audiobook playback later) is actively
-    // playing. When set, the reader yields the volume keys to system volume control.
-    @Volatile var isAudioPlaying: Boolean = false
-}
+// Moved to :feature:reader/commonMain for iOS parity; alias keeps app imports unchanged.
+typealias ReaderStateHolder = com.riffle.feature.reader.ReaderStateHolder
