+++ app/src/main/kotlin/com/riffle/app/feature/reader/VolumeNavigationController.kt
@@ -1,15 +1,4 @@
 package com.riffle.app.feature.reader
 
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.asSharedFlow
-
-
-class VolumeNavigationController constructor() {
-    private val _events = MutableSharedFlow<VolumeNavEvent>(extraBufferCapacity = 1)
-    val events: SharedFlow<VolumeNavEvent> = _events.asSharedFlow()
-
-    fun emit(event: VolumeNavEvent) {
-        _events.tryEmit(event)
-    }
-}
+// Moved to :feature:reader/commonMain for iOS parity; alias keeps app imports unchanged.
+typealias VolumeNavigationController = com.riffle.feature.reader.VolumeNavigationController
