+++ core/data/src/iosMain/kotlin/com/riffle/core/data/di/IosDataKoinModule.kt
@@ -6,6 +6,8 @@ import com.riffle.core.data.IosDeviceIdStoreImpl
 import com.riffle.core.data.IosDeviceLabelResolver
 import com.riffle.core.data.IosFileStore
 import com.riffle.core.data.IosTokenStorage
+import com.riffle.core.data.comic.panel.IosColorPageDecoder
+import com.riffle.core.data.comic.panel.IosPageImageDecoder
 import com.riffle.core.data.localfiles.FolderPickerInterface
 import com.riffle.core.data.localfiles.IosCopyInService
 import com.riffle.core.data.localfiles.IosFolderPicker
@@ -13,10 +15,11 @@ import com.riffle.core.data.localfiles.IosFolderWalker
 import com.riffle.core.data.localfiles.IosLocalFilesFolderRepository
 import com.riffle.core.data.localfiles.IosLocalFilesScanner
 import com.riffle.core.data.localfiles.IosLocalFilesSourceInstaller
-import com.riffle.core.data.comic.panel.IosColorPageDecoder
-import com.riffle.core.data.comic.panel.IosPageImageDecoder
 import com.riffle.core.data.localfiles.LocalFilesInstallerInterface
 import com.riffle.core.domain.ConnectivityObserver
+import com.riffle.core.domain.DeviceIdStore
+import com.riffle.core.domain.DeviceLabelResolver
+import com.riffle.core.domain.TokenStorage
 import com.riffle.core.domain.comic.panel.ColorPageDecoder
 import com.riffle.core.domain.comic.panel.InMemoryPanelStore
 import com.riffle.core.domain.comic.panel.PageImageDecoder
@@ -24,9 +27,6 @@ import com.riffle.core.domain.comic.panel.PanelDetectionConfig
 import com.riffle.core.domain.comic.panel.PanelEngine
 import com.riffle.core.domain.comic.panel.PanelOrchestrator
 import com.riffle.core.domain.comic.panel.PanelStore
-import com.riffle.core.domain.DeviceIdStore
-import com.riffle.core.domain.DeviceLabelResolver
-import com.riffle.core.domain.TokenStorage
 import org.koin.dsl.module
 
 val iosDataModule = module {
