+++ feature/reader/src/commonMain/kotlin/com/riffle/feature/reader/VolumeNavEvent.kt
@@ -1,4 +1,4 @@
-package com.riffle.app.feature.reader
+package com.riffle.feature.reader
 
 sealed class VolumeNavEvent {
     data object Forward : VolumeNavEvent()
