+++ feature/reader/src/commonMain/kotlin/com/riffle/feature/reader/VolumeNavigationController.kt
@@ -1,0 +1,14 @@
+package com.riffle.feature.reader
+
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.asSharedFlow
+
+class VolumeNavigationController constructor() {
+    private val _events = MutableSharedFlow<VolumeNavEvent>(extraBufferCapacity = 1)
+    val events: SharedFlow<VolumeNavEvent> = _events.asSharedFlow()
+
+    fun emit(event: VolumeNavEvent) {
+        _events.tryEmit(event)
+    }
+}
